+++ app/src/main/java/com/heyanle/closure/net/model/GameLoginReq.kt
@@ -6,5 +6,5 @@ package com.heyanle.closure.net.model
  */
 data class GameLoginReq (
     var account: String,
-    var platform: Int,
+    var platform: Long,
 )
